refactor(clinic-form): integrate custom form hook
