+++ src/app/main/landlord/create-clinic/steps/PropertyProofSection.tsx
@@ -1,9 +1,5 @@
 import StepSectionBase, { StepSectionProps } from "./StepSectionBase";
 
-export interface PropertyProofData {
-  documentFile: File | null;
-}
-
 export default function PropertyProof({
   onClickPrimary,
   onClickSecondary
