--- src/hooks/placeholder.md
@@ -1,9 +1,0 @@
-# Hooks Folder
-
-This folder contains custom React hooks.
-
-Example:
-
-- `useAuth.ts`
-- `useFetch.ts`
-- `useTheme.ts`
